fix dependency check
