+++ tests/tacacs/test_authorization.py
@@ -5,7 +5,8 @@ from _pytest.outcomes import Failed
 import time
 
 from tests.common.helpers.tacacs.tacacs_helper import stop_tacacs_server, start_tacacs_server, \
-    per_command_authorization_skip_versions, remove_all_tacacs_server, get_ld_path
+    per_command_authorization_skip_versions, remove_all_tacacs_server, get_ld_path, \
+    check_tacacs  # noqa: F401
 from tests.tacacs.utils import change_and_wait_aaa_config_update, ensure_tacacs_server_running_after_ut, \
     ssh_connect_remote_retry, ssh_run_command, TIMEOUT_LIMIT, \
     cleanup_tacacs_log, count_authorization_request       # noqa: F401
@@ -158,7 +159,7 @@ def test_authorization_tacacs_only(
                                 enum_rand_one_per_hwsku_hostname,
                                 setup_authorization_tacacs,
                                 tacacs_creds,
-                                check_tacacs,
+                                check_tacacs,  # noqa: F811
                                 remote_user_client,
                                 remote_rw_user_client):
 
@@ -229,7 +230,7 @@ def test_authorization_tacacs_only_some_server_down(
         setup_authorization_tacacs,
         tacacs_creds,
         ptfhost,
-        check_tacacs,
+        check_tacacs,  # noqa: F811
         remote_user_client):
     """
         Setup multiple tacacs server for this UT.
@@ -265,7 +266,7 @@ def test_authorization_tacacs_only_some_server_down(
 
 
 def test_authorization_tacacs_only_then_server_down_after_login(
-        setup_authorization_tacacs, ptfhost, check_tacacs,
+        setup_authorization_tacacs, ptfhost, check_tacacs,  # noqa: F811
         remote_user_client, ensure_tacacs_server_running_after_ut):  # noqa: F811
 
     # Verify when server are accessible, TACACS+ user can run command in server side whitelist.
@@ -288,7 +289,7 @@ def test_authorization_tacacs_only_then_server_down_after_login(
 
 def test_authorization_tacacs_and_local(
         duthosts, enum_rand_one_per_hwsku_hostname,
-        setup_authorization_tacacs_local, tacacs_creds, check_tacacs, remote_user_client):
+        setup_authorization_tacacs_local, tacacs_creds, check_tacacs, remote_user_client):  # noqa: F811
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
 
     """
@@ -355,7 +356,7 @@ def test_authorization_tacacs_and_local_then_server_down_after_login(
 
 def test_authorization_local(
         duthosts, enum_rand_one_per_hwsku_hostname,
-        tacacs_creds, ptfhost, check_tacacs,
+        tacacs_creds, ptfhost, check_tacacs,  # noqa: F811
         remote_user_client, local_user_client, ensure_tacacs_server_running_after_ut):  # noqa: F811
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
 
@@ -392,7 +393,7 @@ def test_authorization_local(
 
 def test_bypass_authorization(
         duthosts, enum_rand_one_per_hwsku_hostname,
-        setup_authorization_tacacs, check_tacacs, remote_user_client):
+        setup_authorization_tacacs, check_tacacs, remote_user_client):  # noqa: F811
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
 
     """
@@ -450,7 +451,7 @@ def test_bypass_authorization(
 
 def test_backward_compatibility_disable_authorization(
         duthosts, enum_rand_one_per_hwsku_hostname,
-        tacacs_creds, ptfhost, check_tacacs,
+        tacacs_creds, ptfhost, check_tacacs,  # noqa: F811
         remote_user_client, local_user_client, ensure_tacacs_server_running_after_ut):  # noqa: F811
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
 
@@ -500,7 +501,7 @@ def test_tacacs_authorization_wildcard(
                                     enum_rand_one_per_hwsku_hostname,
                                     setup_authorization_tacacs,
                                     tacacs_creds,
-                                    check_tacacs,
+                                    check_tacacs,  # noqa: F811
                                     remote_user_client,
                                     remote_rw_user_client):
     # Create files for command with wildcards
@@ -551,7 +552,7 @@ def test_stop_request_next_server_after_reject(
                                             setup_authorization_tacacs,
                                             tacacs_creds,
                                             ptfhost,
-                                            check_tacacs):
+                                            check_tacacs):  # noqa: F811
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
 
     # not ignore on version >= 202305
@@ -596,7 +597,7 @@ def test_fallback_to_local_authorization_with_config_reload(
                                     enum_rand_one_per_hwsku_hostname,
                                     setup_authorization_tacacs,
                                     tacacs_creds,
-                                    check_tacacs,
+                                    check_tacacs,  # noqa: F811
                                     remote_user_client,
                                     remote_rw_user_client):
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
@@ -652,7 +653,7 @@ def test_tacacs_authorization_commands_during_login(
                                                 enum_rand_one_per_hwsku_hostname,
                                                 setup_authorization_tacacs,
                                                 tacacs_creds,
-                                                check_tacacs,
+                                                check_tacacs,  # noqa: F811
                                                 remote_rw_user_client):
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
     duthost.shell("sudo config aaa authentication debug disable")
